pyproject.toml
[project]
name = "sushi-taisho"
version = "0.1.0"
description = "Sushi-Bot: Autonomous Sushi Picking Robot for AMD Open Robotics Hackathon 2025"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "ruff>=0.1.0",
    # TODO: 実行に必要な依存関係を追記
    # "lerobot",
    # "numpy",
    # "opencv-python",
    # その他の必要な依存関係
    "ml-networks",
    "einops>=0.8.1",
    "numpy>=1.24.4",
    "torchcfm>=1.0.5",
    "timm>=1.0.22",
    "torch-ema>=0.3",
    "diffusers>=0.35.2",
    "wandb<0.22.0",
    "rich>=14.2.0",
    "hydra-core>=1.3.2",
    "lerobot",
]

[project.scripts]
train = "mission2.code.scripts.train:main"
infer = "mission2.code.scripts.inference:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mission2"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["mission2"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv.sources]
ml-networks = { git = "https://github.com/keio-crl/ml-networks.git" }
lerobot = { path = "third_party/lerobot" }
